pyproject.toml
[project]
name = "sibadi-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiogram-dialog>=2.4.0",
    "cachetools==5.*",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "types-cachetools>=6.2.0.20251022",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.11",
    "types-requests>=2.32.4.20250913",
    "wemake-python-styleguide>=1.4.0",
]
